+++ packages/kiarina/CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Removed
+- Removed `kiarina-lib-google-cloud-storage` from dependencies
+
 ## [1.37.0] - 2026-05-27
 
 ### Added