+++ packages/kiarina/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.0.0] - 2026-06-10
+
 ### Removed
 - Removed `kiarina-lib-atlassian` from dependencies
 - Removed `kiarina-llm` from dependencies